Cleaned up CA1821 (Remove empty finalizers) and fixed a bug in the fixer.
The fixer removed the squiggled node and for VB that would be a MethodStatementSyntax instead of a MethodBlockSyntax and the RemoveNode call fails. Fixing it by simply calling SyntaxGenerator.GetDeclartation before removing and that does thte right thing.

Fixes #2257
